- fix issue LIB-105, mark as read available when message is already read
modified:   Airship/UI/Default/Inbox/Classes/Shared/UAInboxMessageListController.m

added two private methods:
- (NSArray*)selectedMessagesFromSet:(NSSet*)set;
- (NSDictionary*)readUnreadCountFromMessages:(NSArray*)messages;

Added logic to the - (void)refreshBatchUpdateButtons  method to
modifiy the presentation of the moveItems button, which is really a markAsRead button.
During the execution of the method,  the table view cells selected are matched to
their respective UAInboxMessage, each group of selected messages are parsed to
identify read/unread status. The accesibility and display count of the markAsRead button
are then determined from the information.

-  modified:   InboxSample/Classes/Shared/InboxSampleAppDelegate.m

Added #import "UAAnalytics.h" to InboxSampleAppDelegate.h to get the project to compile