- refactored initial pass at LIB-105,  mark as read available when message is already read
Previous solution  created an NSDictionary, and  was deemed to heavy to scale for a large amount
of messages after review. Current implementation contains more difficult program logic, but maintains
a smaller memory footprint, and allocates less objects to the autorelease pool.

  modified:   Airship/UI/Default/Inbox/Classes/Shared/UAInboxMessageListController.h
  modified:   Airship/UI/Default/Inbox/Classes/Shared/UAInboxMessageListController.m

  Added ivar:
  NSMutableSet *setOfUnreadMessagesInSelection

  Added the following methods:

- (UAInboxMessage *)messageForIndexPath:(NSIndexPath *)indexPath;
- (void)updateSetOfUnreadMessagesWithMessage:(UAInboxMessage *)message atIndexPath:(NSIndexPath *)indexPath;
- (NSIndexPath *)checkSetOfIndexPaths:(NSSet *)setOfPaths forIndexPath:(NSIndexPath *)indexPath;
- (NSUInteger) countOfUnreadMessagesInSetOfIndexPaths:(NSSet *)set;

A running total of unread messages in the current set of selected messages in maintained at each call of
didSelectRowAtIndexPath, this is referenced for UI updates.
Tested with instruments, went through several iterations of message selection and deletion
for both modal and navigation views.