- moving forward with UALocationManager
moved the UALocationManagerDelegate to UALocationServicesDelegate, modified
it for use with both UALocationManager and UASingleLocationAcquireAndUpload

working on automaticStandardLocation services, plan to fire the UASingleLocationAquireAndUpload
on app start, and start a timed service.

moved status enums to a separate file to be used across classes. Added a UALocationServices.h
file to contain common parameters. Renamed parameters and enums

separated CLLocationManager+Test catagory to a separate class to share with UASingleLocationAcquireAndUpload
test class

updated specs