MB-2186: Better UI when executing Message Center actions (#1227)
Improve UI for message center actions by NOT waiting until message has been fetched from the cloud before calling the inbox delegate. Now, if the message is not available on the device, the messageID is passed to the inbox delegate. This allows the inbox delegate to almost immediately update the UI and "do the right thing", e.g. switch to a message view and show loading indicators.
 * UAInboxDelegate method showInboxMessage: has been deprecated. The new method is: showMessageForID:
 * UADefaultMessageCenter methods displayMessage:animated: and displayMessage: have been deprecated. The new methods are: displayMessageForID:animated: and displayMessageForID:
 * UADefaultMessageCenterListViewController methods displayMessage: and displayMessage:onError: have been deprecated. The new methods are: displayMessageForID: and displayMessageForID:onError:
 * UAMessageCenterMessageViewProtocol method loadMessage:onlyIfChange: has been deprecated. The new method is loadMessageForID:onlyIfChanged:onError: